requirements.txt
Django==4.1.7
behave-django==1.4.0
